Support buidling on Ubuntu
Ubuntu stores overlays under `/boot/firmware/overlays`, instead of just
`/boot/overlays` like many other operating systems for the Pi.  The
location of the folder where `overlays` can be found is defined in
`/etc/default/rpi-eeprom-update`.

This change checks that file for `BOOTFS`, and defaults to `/boot` if
not defined there.