Support 5.13.y Kernels
This builds just fine on 5.13 (and probably 5.11 and 5.12 too).

Ubuntu 21.10, which is the suggested version of Ubuntu Server to install
for the CM4, runs on 5.13.

This does not, however, seem to properly install.  DKMS will build it,
and place it under
`/lib/modules/5.13.0-1010-raspi/updates/dkms/emc2301.ko`, but not under
`/lib/modules/5.13.0-1010-raspi/kernel/drivers/hwmon/`.  If we manually
load the module with `insmod`, the fan is properly controlled.